Use new file organisation component
