+++ comp/process/bundle.go
@@ -14,19 +14,19 @@ package process
 import (
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/process/apiserver"
-	"github.com/DataDog/datadog-agent/comp/process/connectionscheck"
-	"github.com/DataDog/datadog-agent/comp/process/containercheck"
-	"github.com/DataDog/datadog-agent/comp/process/expvars"
+	connectionscheck "github.com/DataDog/datadog-agent/comp/process/connectionscheck/impl"
+	containercheck "github.com/DataDog/datadog-agent/comp/process/containercheck/impl"
+	expvars "github.com/DataDog/datadog-agent/comp/process/expvars/impl"
 	"github.com/DataDog/datadog-agent/comp/process/forwarders"
-	"github.com/DataDog/datadog-agent/comp/process/hostinfo"
-	"github.com/DataDog/datadog-agent/comp/process/podcheck"
-	"github.com/DataDog/datadog-agent/comp/process/processcheck"
-	"github.com/DataDog/datadog-agent/comp/process/processdiscoverycheck"
-	"github.com/DataDog/datadog-agent/comp/process/processeventscheck"
-	"github.com/DataDog/datadog-agent/comp/process/profiler"
-	"github.com/DataDog/datadog-agent/comp/process/rtcontainercheck"
-	"github.com/DataDog/datadog-agent/comp/process/runner"
-	"github.com/DataDog/datadog-agent/comp/process/submitter"
+	hostinfo "github.com/DataDog/datadog-agent/comp/process/hostinfo/impl"
+	podcheck "github.com/DataDog/datadog-agent/comp/process/podcheck/impl"
+	processcheck "github.com/DataDog/datadog-agent/comp/process/processcheck/impl"
+	processdiscoverycheck "github.com/DataDog/datadog-agent/comp/process/processdiscoverycheck/impl"
+	processeventscheck "github.com/DataDog/datadog-agent/comp/process/processeventscheck/impl"
+	profiler "github.com/DataDog/datadog-agent/comp/process/profiler/impl"
+	rtcontainercheck "github.com/DataDog/datadog-agent/comp/process/rtcontainercheck/impl"
+	runner "github.com/DataDog/datadog-agent/comp/process/runner/impl"
+	submitter "github.com/DataDog/datadog-agent/comp/process/submitter/impl"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
