+++ comp/process/connectionscheck/impl/check.go
@@ -3,15 +3,23 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package connectionscheck
+// Package impl implements a component to handle Connections data collection in the Process Agent.
+package impl
 
 import (
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
+	"github.com/DataDog/datadog-agent/comp/process/connectionscheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+var Module = fxutil.Component(
+	fx.Provide(newCheck),
 )
 
 var _ types.CheckComponent = (*check)(nil)
@@ -31,7 +39,7 @@ type result struct {
 	fx.Out
 
 	Check     types.ProvidesCheck
-	Component Component
+	Component connectionscheck.Component
 }
 
 func newCheck(deps dependencies) result {
