+++ comp/process/podcheck/impl/check.go
@@ -3,13 +3,21 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package podcheck
+// Package impl implements a component to handle Kubernetes data collection in the Process Agent.
+package impl
 
 import (
 	"go.uber.org/fx"
 
+	"github.com/DataDog/datadog-agent/comp/process/podcheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+var Module = fxutil.Component(
+	fx.Provide(newCheck),
 )
 
 var _ types.CheckComponent = (*check)(nil)
@@ -22,7 +30,7 @@ type result struct {
 	fx.Out
 
 	Check     types.ProvidesCheck
-	Component Component
+	Component podcheck.Component
 }
 
 func newCheck() result {
