+++ comp/process/processcheck/component.go
@@ -7,10 +7,7 @@
 package processcheck
 
 import (
-	"go.uber.org/fx"
-
 	"github.com/DataDog/datadog-agent/comp/process/types"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
 // team: processes
@@ -18,12 +15,3 @@ import (
 type Component interface {
 	types.CheckComponent
 }
-
-// Module defines the fx options for this component.
-var Module = fxutil.Component(
-	fx.Provide(newCheck),
-)
-
-var MockModule = fxutil.Component(
-	fx.Provide(newMock),
-)
