+++ comp/process/processeventscheck/impl/check.go
@@ -3,20 +3,28 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package containercheck
+// Package impl implements a component to handle Process Events data collection in the Process Agent.
+package impl
 
 import (
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
+	"github.com/DataDog/datadog-agent/comp/process/processeventscheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+var Module = fxutil.Component(
+	fx.Provide(newCheck),
 )
 
 var _ types.CheckComponent = (*check)(nil)
 
 type check struct {
-	containerCheck *checks.ContainerCheck
+	processEventsCheck *checks.ProcessEventsCheck
 }
 
 type dependencies struct {
@@ -29,12 +37,12 @@ type result struct {
 	fx.Out
 
 	Check     types.ProvidesCheck
-	Component Component
+	Component processeventscheck.Component
 }
 
 func newCheck(deps dependencies) result {
 	c := &check{
-		containerCheck: checks.NewContainerCheck(deps.Config),
+		processEventsCheck: checks.NewProcessEventsCheck(deps.Config),
 	}
 	return result{
 		Check: types.ProvidesCheck{
@@ -45,5 +53,5 @@ func newCheck(deps dependencies) result {
 }
 
 func (c *check) Object() checks.Check {
-	return c.containerCheck
+	return c.processEventsCheck
 }
