+++ comp/process/profiler/component.go
@@ -6,18 +6,7 @@
 // Package profiler implements a component to handle starting and stopping the internal profiler.
 package profiler
 
-import (
-	"go.uber.org/fx"
-
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
 // team: processes
 
 type Component interface {
 }
-
-// Module defines the fx options for this component.
-var Module = fxutil.Component(
-	fx.Provide(newProfiler),
-)
