+++ comp/process/rtcontainercheck/impl/check.go
@@ -3,20 +3,28 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package processeventscheck
+// Package impl implements a component to handle realtime Container data collection in the Process Agent.
+package impl
 
 import (
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
+	"github.com/DataDog/datadog-agent/comp/process/rtcontainercheck"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+var Module = fxutil.Component(
+	fx.Provide(newCheck),
 )
 
 var _ types.CheckComponent = (*check)(nil)
 
 type check struct {
-	processEventsCheck *checks.ProcessEventsCheck
+	rtContainerCheck *checks.RTContainerCheck
 }
 
 type dependencies struct {
@@ -29,12 +37,12 @@ type result struct {
 	fx.Out
 
 	Check     types.ProvidesCheck
-	Component Component
+	Component rtcontainercheck.Component
 }
 
 func newCheck(deps dependencies) result {
 	c := &check{
-		processEventsCheck: checks.NewProcessEventsCheck(deps.Config),
+		rtContainerCheck: checks.NewRTContainerCheck(deps.Config),
 	}
 	return result{
 		Check: types.ProvidesCheck{
@@ -45,5 +53,5 @@ func newCheck(deps dependencies) result {
 }
 
 func (c *check) Object() checks.Check {
-	return c.processEventsCheck
+	return c.rtContainerCheck
 }
