+++ comp/process/submitter/impl/submitter_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package submitter
+package impl
 
 import (
 	"testing"
@@ -12,12 +12,13 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/process/forwarders"
-	"github.com/DataDog/datadog-agent/comp/process/hostinfo"
+	hostinfo "github.com/DataDog/datadog-agent/comp/process/hostinfo/impl"
+	submitterComp "github.com/DataDog/datadog-agent/comp/process/submitter"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
 func TestSubmitterLifecycle(t *testing.T) {
-	_ = fxutil.Test[Component](t, fx.Options(
+	_ = fxutil.Test[submitterComp.Component](t, fx.Options(
 		hostinfo.MockModule,
 		core.MockBundle,
 		forwarders.MockModule,
